dnsio_tcp: buffered reads
Before this change, we had fair connection balance (each live conn
can only complete at most one transaction per loop iteration), but
we refused to buffer beyond the first available request in user
space, and thus commonly required 2x recv() calls per request (to
fetch req length and then fetch the exact amount of data for that
one req).

With this change fair connection balance is maintained and we
buffer incoming traffic in user space. This avoids the inefficient
double-recv() pattern and make handling bursts of pipelined
requests much more efficient as well.

The receive buffer is 4K in size.  Any recv() operation attempts
to fill the whole buffer (if TCP has the data).  So long as one or
more whole requests are in our buffer, we process one per
eventloop iteration (and make the eventloop not block for lack of
natural events temporarily) and do not poll for new network data
on this connection until they're all handled.  Any remaining
partial request is moved to the bottom of the buffer before
flipping back to recv() operations again.

The PROXY protocol parser was also switched over to share the same
read buffering system as well, and efficiently handles the case
where one or more requests are already available in our buffer
after parsing the PROXY header from the initial full-sized recv().

+++ src/dnsio_tcp.h
@@ -26,7 +26,7 @@
 #include <gdnsd/net.h>
 
 // Called before threads are started, to initialize a thread registry used for stopping
-void dnsio_tcp_init(unsigned num_threads);
+void dnsio_tcp_init(size_t num_threads);
 
 // Uses the above registry to ask threads to stop.  They will exit when
 // outstanding TCP connections drain due to close/timeout.
